requirements.txt
fastapi
uvicorn
tensorflow
numpy
pillow
python-multipart  # for handling image uploads
